+++ src/interface/interfaceconfig.py
@@ -1,0 +1,5 @@
+import libtcodpy as libtcod
+
+MOUSE = libtcod.Mouse()
+KEY = libtcod.Key()
+
