+++ src/libtcod.cfg
@@ -3,7 +3,7 @@
 	overriding what's defined in the sample/game code.
 	The main goal of this file is to allow the player
 	to tweak the game in case it doesn't work properly
-	on his system and there are no options screen/config
+	on his system and there are no options screen/gameconfig
     files available in the game.
 */
 libtcod {
